+++ mock/user.ts
@@ -1,8 +1,8 @@
 import { Request, Response } from 'express';
 
-function getFakeCaptcha(req: Request, res: Response) {
-  return res.json('captcha-xxx');
-}
+// function getFakeCaptcha(req: Request, res: Response) {
+//   return 123456;
+// }
 // 代码中会兼容本地 service mock 以及部署站点的静态数据
 export default {
   // 支持值为 Object 和 Array
@@ -141,5 +141,5 @@ export default {
     });
   },
 
-  'GET  /api/login/captcha': getFakeCaptcha,
+  'GET  /api/login/captcha': { data: { captcha: 123456 } },
 };