+++ src/pages/Authorized.tsx
@@ -21,7 +21,7 @@ const AuthComponent: React.FC<AuthComponentProps> = ({
 }) => {
   const { currentUser } = user;
   const { routes = [] } = route;
-  const isLogin = currentUser && currentUser.name;
+  const isLogin = currentUser && currentUser.manage_name;
   return (
     <Authorized
       authority={getRouteAuthority(location.pathname, routes) || ''}