+++ src/services/login.ts
@@ -15,5 +15,10 @@ export async function fakeAccountLogin(params: LoginParamsType) {
 }
 
 export async function getFakeCaptcha(mobile: string) {
+  // return api.get('/pc/manage/getFakeCaptcha', {
+  //   params: {
+  //     mobile: mobile,
+  //   },
+  // });
   return api(`/api/login/captcha?mobile=${mobile}`);
 }